requirements.txt
cycler==0.10.0
decorator==4.3.0
kiwisolver==1.0.1
numpy==1.15.4
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.0
python-dateutil==2.7.5
six==1.12.0
